+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/service/IdmProfileService.java
@@ -4,15 +4,17 @@ import java.io.Serializable;
 
 import org.springframework.web.multipart.MultipartFile;
 
+import eu.bcvsolutions.idm.core.api.config.domain.PrivateIdentityConfiguration;
 import eu.bcvsolutions.idm.core.api.dto.IdmProfileDto;
 import eu.bcvsolutions.idm.core.api.dto.filter.IdmProfileFilter;
 import eu.bcvsolutions.idm.core.api.exception.ForbiddenEntityException;
+import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
 import eu.bcvsolutions.idm.core.api.script.ScriptEnabled;
 import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 import eu.bcvsolutions.idm.core.security.api.service.AuthorizableService;
 
 /**
- * Operations with profiles
+ * Operations with profiles.
  * 
  * @author Radek Tomiška
  * @since 9.0.0
@@ -47,11 +49,13 @@ public interface IdmProfileService extends
 	 * Upload new image version for the given profile
 	 * 
 	 * @param profile persisted profile
-	 * @param data
-	 * @param fileName
+	 * @param data - one of image/* content type is required.
+	 * @param fileName Original filer name 
 	 * @param permission permissions to evaluate (AND)
 	 * @return
 	 * @throws ForbiddenEntityException if authorization policies doesn't met
+	 * @throws ResultCodeException if content type is different than image (one of  image/*)
+	 * @throws ResultCodeException if content size is higher than configured maximum (by {@link PrivateIdentityConfiguration#PROPERTY_IDENTITY_PROFILE_IMAGE_MAX_FILE_SIZE})
 	 */
 	IdmProfileDto uploadImage(IdmProfileDto profile, MultipartFile data, String fileName, BasePermission... permission);